pyproject.toml
[project]
name = "genai-launchpad"
version = "1.0.0"
description = "GenAI Launchpad by Datalumina"
requires-python = ">=3.13.7"
dependencies = [
    "alembic>=1.17.2",
    "celery>=5.6.0",
    "fastapi>=0.123.8",
    "jinja2>=3.1.6",
    "langfuse>=3.10.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.26",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "redis>=7.1.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "graphviz>=0.21",
    "ipykernel>=7.1",
    "ipython>=9.8.0",
    "nest-asyncio>=1.6.0",
    "ruff>=0.14.8",
]

[tool.ruff]
exclude = [".venv", "app/core/commands", "app/alembic", ".sh"]
